% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countries.R
\name{aq_countries}
\alias{aq_countries}
\title{Providing a simple listing of countries within the platform.}
\usage{
aq_countries(limit = 100, page = 1)
}
\arguments{
\item{limit}{Change the number of results returned, max is 1000.}

\item{page}{The page of the results to query. This can be useful if e.g. there are 2000 measurements, then first use page=1 and page=2 with limit=100 to get all measurements for your query.}
}
\value{
data.frame (dplyr "tbl_df") with 3 columns:
\itemize{
\item the number of measures for a country ("count"),
\item the ISO 3166-1 alpha-2 code of the country ("code"),
\item and its name ("name").
}
}
\description{
Providing a simple listing of countries within the platform.
}
\details{
For queries involving a country argument,
the code of the country, not its name, should be used.
}
\examples{
aq_countries()
}

