% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latest.R
\name{aq_latest}
\alias{aq_latest}
\title{Provides the latest value of each available parameter for every location in the system.}
\usage{
aq_latest(country = NULL, city = NULL, location = NULL,
  parameter = NULL, has_geo = NULL, value_from = NULL, value_to = NULL)
}
\arguments{
\item{country}{Limit results by a certain country -- a two-letters codem see countries() for finding code based on name.}

\item{city}{Limit results by a certain city.}

\item{location}{Limit results by a certain location.}

\item{parameter}{Limit to only a certain parameter (valid values are 'pm25', 'pm10', 'so2', 'no2', 'o3', 'co' and 'bc').
If no parameter is given, all parameters are retrieved.}

\item{has_geo}{has_geo Filter out items that have or do not have geographic information.}

\item{value_from}{Show results above value threshold, useful in combination with \code{parameter}.}

\item{value_to}{Show results below value threshold, useful in combination with \code{parameter}.}
}
\value{
a data.frame (dplyr "tbl_df") with 11 columns:
\itemize{
 \item the name of the location ("location"),
 \item the city it is in ("city"),
 \item the code of country it is in ("country"),
 \item its longitude ("longitude") and latitude if available ("latitude"),
 \item the parameter ("parameter")
 \item the value of the measurement ("value")
 \item the last time and date at which the value was updated ("lastUpdated"),
 \item the unit of the measure ("unit")
 \item and finally an URL encoded version of the city name ("cityURL")
 \item and of the location name ("locationURL").
}.
}
\description{
Provides the latest value of each available parameter for every location in the system.
}
\details{
For queries involving a city or location argument,
the URL-encoded name of the city/location (as in cityURL/locationURL),
not its name, should be used.
 You can query any nested combination of country/location/city (level 1, 2 and 3),
 with only one value for each argument.
  If you write inconsistent combination such as city="Paris" and country="IN", an error message will be returned.
  If you write city="Delhi", you do not need to write the code of the country, unless
  one day there is a city with the same name in another country.
}
\examples{
\dontrun{
aq_latest(country='IN', city='Chennai')
aq_latest(parameter='co')
}
}

