% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cities.R
\name{cities}
\alias{cities}
\title{Providing a simple listing of cities within the platform.}
\usage{
cities(country = NULL)
}
\arguments{
\item{country}{}
}
\value{
a data.table with locations, count, country, city columns.
}
\description{
Providing a simple listing of cities within the platform.
}
\examples{
cities(country="IN")
}

