% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countries.R
\name{countries}
\alias{countries}
\title{Providing a simple listing of countries within the platform.}
\usage{
countries()
}
\value{
data.frame (dplyr "tbl_df") with code and name for each country,
as well as the number of sources in each country.
}
\description{
Providing a simple listing of countries within the platform.
}
\examples{
countries()
}

