% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countries.R
\name{aq_countries}
\alias{aq_countries}
\title{Provides a simple listing of countries within the platform.}
\usage{
aq_countries(limit = 10000, page = NULL)
}
\arguments{
\item{limit}{Change the number of results returned, max is 10000 by page.
If no \code{page} value, then all results or \code{limit} * 100 (maximal number of pages)
is returned.}

\item{page}{The page of the results to query (max 100).}
}
\value{
A results data.frame (dplyr "tbl_df") with 3 columns:
\itemize{
\item the number of measures for a country ("count"),
\item the ISO 3166-1 alpha-2 code of the country ("code"),
\item and its name ("name").
}
and 2 attributes, a meta data.frame (dplyr "tbl_df") with 1 line and 5 columns:
\itemize{
\item the API name ("name"),
\item the license of the data ("license"),
\item the website url ("website"),
\item the queried page ("page"),
\item the limit on the number of results ("limit"),
\item the number of results found on the platform for the query ("found")
}
and a timestamp data.frame (dplyr "tbl_df") with the query time and the last time at which the data was modified on the platform.
}
\description{
Provides a simple listing of countries within the platform.
}
\details{
For queries involving a country argument,
the code of the country, not its name, should be used.
}
\examples{
\dontrun{
countries <- aq_countries()
countries
}
}
