% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cities.R
\name{cities}
\alias{cities}
\title{Providing a simple listing of cities within the platform.}
\usage{
cities(country = NULL)
}
\arguments{
\item{country}{Limit results by a certain country -- a two-letters codem see countries() for finding code based on name.}
}
\value{
a data.frame (dplyr "tbl_df") with locations, count, country, city columns,
and also an URL encoded string for the city.
}
\description{
Providing a simple listing of cities within the platform.
}
\details{
Please note that if an argument is composed by several words,
e.g. 'RK Puram' as a location, it has to be written 'RK+Puram' as in a URL.
}
\examples{
cities(country='IN')
}

