% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countries.R
\name{countries}
\alias{countries}
\title{Providing a simple listing of countries within the platform.}
\usage{
countries()
}
\value{
data.frame (dplyr "tbl_df") with 3 columns:
\itemize{
\item the number of measures for a country ("count"),
\item the code of the country ("code"),
\item and its name ("name").
}
}
\description{
Providing a simple listing of countries within the platform.
}
\details{
For queries involving a country argument,
the code of the country, not its name, should be used.
}
\examples{
countries()
}

