% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/countries.R
\name{countries}
\alias{countries}
\title{Providing a simple listing of countries within the platform.}
\usage{
countries()
}
\description{
Providing a simple listing of countries within the platform.
}
\examples{
countries()
}

