% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/latest.R
\name{latest}
\alias{latest}
\title{Provides the latest value of each available parameter for every location in the system.}
\usage{
latest(country = NULL, city = NULL, location = NULL, parameter = NULL,
  has_geo = NULL, value_from = NULL, value_to = NULL)
}
\arguments{
\item{country}{Limit results by a certain country -- a two-letters codem see countries() for finding code based on name.}

\item{city}{Limit results by a certain city.}

\item{location}{Limit results by a certain location.}

\item{parameter}{Limit to only a certain parameter (valid values are 'pm25', 'pm10', 'so2', 'no2', 'o3', 'co' and 'bc').
If no parameter is given, all parameters are retrieved.}

\item{has_geo}{has_geo Filter out items that have or do not have geographic information.}

\item{value_from}{Show results above value threshold, useful in combination with \code{parameter}.}

\item{value_to}{Show results below value threshold, useful in combination with \code{parameter}.}
}
\value{
a data.frame (dplyr "tbl_df") with location, city, country, parameter, value, the last time at which this value was updated, unit, latitude and longitude (NA when non available).
}
\description{
Provides the latest value of each available parameter for every location in the system.
}
\details{
Please note that if an argument is composed by several words, e.g. 'RK Puram' as a location, it has to be written 'RK+Puram' as in a URL.
 You can query any combination of country/location/city, but only one of them at the same time.
  If you write inconsistent combination such as city="Paris" and country="IN", an error message will be returned.
}
\examples{
\dontrun{
latest(country='IN', city='Chennai')
latest(parameter='co')
}
}

