% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locations.R
\name{locations}
\alias{locations}
\title{Providing data about distinct measurement locations.}
\usage{
locations(city = NULL, country = NULL, location = NULL,
  parameter = NULL, has_geo = NULL, value_from = NULL, value_to = NULL,
  date_from = NULL, date_to = NULL)
}
\arguments{
\item{city}{Limit results by a certain city.}

\item{country}{Limit results by a certain country.}

\item{location}{Limit results by a certain location.}

\item{parameter}{Limit to only a certain parameter (valid values are 'pm25', 'pm10', 'so2', 'no2', 'o3', 'co' and 'bc').
If no parameter is given, all parameters are retrieved.}

\item{has_geo}{Filter out items that have or do not have geographic information.Can only be TRUE for now.}

\item{value_from}{Show results above value threshold, useful in combination with \code{parameter}.}

\item{value_to}{Show results below value threshold, useful in combination with \code{parameter}.}

\item{date_from}{Show results after a certain date. (ex. '2015-12-20').}

\item{date_to}{Show results before a certain date. (ex. '2015-12-20')}
}
\value{
A data.table with for each location: its name, the city and the country it is in, the number of measures for this location,
 the name of the source, and the dates and times at which it was first and last updated, respectively. City and location are
 also provided as URL encoded strings which makes further queries easier.
 @details Please note that if an argument is composed by several words, e.g. 'RK Puram' as a location, it has to be written 'RK+Puram' as in a URL.
}
\description{
Providing data about distinct measurement locations.
}
\examples{
locations(country='IN')
locations(city='Houston', parameter='co')
}

