% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/locations.R
\name{locations}
\alias{locations}
\title{Providing data about distinct measurement locations.}
\usage{
locations(city = NULL, country = NULL, location = NULL,
  parameter = NULL, has_geo = NULL, value_from = NULL, value_to = NULL,
  date_from = NULL, date_to = NULL)
}
\arguments{
\item{date_to}{}
}
\value{
A data.table with for each location: its name, the city and the country it is in, the number of measures for this location,
 the name of the source, and the dates and times at which it was first and last updated, respectively.
}
\description{
Providing data about distinct measurement locations.
}
\examples{
locations(country="IN")
}

