% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ropenaq-package.R
\docType{package}
\name{ropenaq-package}
\alias{ropenaq}
\alias{ropenaq-package}
\title{ropenaq: Accesses Air Quality Data from the Open Data Platform OpenAQ}
\description{
Allows access to air quality data from the API of the OpenAQ
    platform <https://docs.openaq.org/>, with the different services the API offers
    (getting measurements for a given query, getting latest measurements, getting
    lists of available countries/cities/locations). This package has been peer-reviewed by rOpenSci (v. 0.1.0).
}
\seealso{
Useful links:
\itemize{
  \item \url{https://docs.ropensci.org/ropenaq}
  \item \url{http://github.com/ropensci/ropenaq}
  \item Report bugs at \url{http://github.com/ropensci/ropenaq/issues}
}

}
\author{
\strong{Maintainer}: Maëlle Salmon \email{maelle.salmon@yahoo.se}

Other contributors:
\itemize{
  \item Andrew MacDonald (Andrew MacDonald reviewed the package for rOpenSci, see https://github.com/ropensci/onboarding/issues/24) [reviewer]
  \item Andy  Teucher (Andy Teucher reviewed the package for rOpenSci, see https://github.com/ropensci/onboarding/issues/24) [reviewer]
  \item Mikayla  Murphy [contributor]
}

}
\keyword{internal}
